refactor(reports): move metadata and result class
move metadata and result class to reports/models.py
