+++ workers/ohsome_quality_analyst/reports/base.py
@@ -2,8 +2,8 @@ from __future__ import annotations  # superfluous in Python 3.10
 
 import logging
 from abc import ABCMeta, abstractmethod
-from dataclasses import asdict, dataclass
-from typing import List, Literal, NamedTuple, Tuple
+from dataclasses import asdict
+from typing import List, NamedTuple, Tuple
 
 import numpy as np
 from geojson import Feature
@@ -14,30 +14,8 @@ from ohsome_quality_analyst.html_templates.template import (
     get_traffic_light,
 )
 from ohsome_quality_analyst.indicators.base import BaseIndicator
+from ohsome_quality_analyst.reports.models import Metadata, Result
 from ohsome_quality_analyst.utils.helper import flatten_dict
-
-
-@dataclass
-class Metadata:
-    """Metadata of a report as defined in the metadata.yaml file"""
-
-    name: str
-    description: str
-    label_description: dict
-
-
-@dataclass
-class Result:
-    """The result of the Report."""
-
-    class_: Literal[1, 2, 3, 4, 5] | None = None
-    description: str = ""
-    html: str = ""
-
-    @property
-    def label(self) -> Literal["green", "yellow", "red", "undefined"]:
-        labels = {1: "red", 2: "yellow", 3: "yellow", 4: "green", 5: "green"}
-        return labels.get(self.class_, "undefined")
 
 
 class IndicatorTopic(NamedTuple):
